+++ diff_test.go
@@ -21,35 +21,44 @@ func TestDiffRoundTrip(t *testing.T) {
 	db := db.NewMemDB()
 	tree, err := NewMutableTree(db, 0, true)
 	require.NoError(t, err)
-	for i := range changeSets {
-		v, err := tree.SaveChangeSet(changeSets[i])
+	for _, cs := range changeSets {
+		for _, pair := range cs.Pairs {
+			if pair.Delete {
+				_, removed, err := tree.Remove(pair.Key)
+				require.True(t, removed)
+				require.NoError(t, err)
+			} else {
+				_, err := tree.Set(pair.Key, pair.Value)
+				require.NoError(t, err)
+			}
+		}
+		_, _, err := tree.SaveVersion()
 		require.NoError(t, err)
-		require.Equal(t, int64(i+1), v)
 	}
 
 	// extract change sets from db
-	var extractChangeSets []*ChangeSet
+	var extractChangeSets []ChangeSet
 	tree2 := NewImmutableTree(db, 0, true)
-	err = tree2.ndb.traverseStateChanges(0, math.MaxInt64, func(version int64, changeSet *ChangeSet) error {
-		extractChangeSets = append(extractChangeSets, changeSet)
+	err = tree2.TraverseStateChanges(0, math.MaxInt64, func(version int64, changeSet *ChangeSet) error {
+		extractChangeSets = append(extractChangeSets, *changeSet)
 		return nil
 	})
 	require.NoError(t, err)
 	require.Equal(t, changeSets, extractChangeSets)
 }
 
-func genChangeSets(r *rand.Rand, n int) []*ChangeSet {
-	var changeSets []*ChangeSet
+func genChangeSets(r *rand.Rand, n int) []ChangeSet {
+	var changeSets []ChangeSet
 
 	for i := 0; i < n; i++ {
-		items := make(map[string]*KVPair)
+		items := make(map[string]KVPair)
 		start, count, step := r.Int63n(1000), r.Int63n(1000), r.Int63n(10)
 		for i := start; i < start+count*step; i += step {
 			value := make([]byte, 8)
 			binary.LittleEndian.PutUint64(value, uint64(i))
 
 			key := fmt.Sprintf("test-%d", i)
-			items[key] = &KVPair{
+			items[key] = KVPair{
 				Key:   []byte(key),
 				Value: value,
 			}
@@ -65,7 +74,7 @@ func genChangeSets(r *rand.Rand, n int) []*ChangeSet {
 				if pair.Delete {
 					continue
 				}
-				items[string(pair.Key)] = &KVPair{
+				items[string(pair.Key)] = KVPair{
 					Key:    pair.Key,
 					Delete: true,
 				}
@@ -77,7 +86,7 @@ func genChangeSets(r *rand.Rand, n int) []*ChangeSet {
 				i := r.Int63n(int64(len(lastChangeSet.Pairs)))
 				pair := lastChangeSet.Pairs[i]
 				if !pair.Delete {
-					items[string(pair.Key)] = &KVPair{
+					items[string(pair.Key)] = KVPair{
 						Key:   pair.Key,
 						Value: pair.Value,
 					}
@@ -93,10 +102,11 @@ func genChangeSets(r *rand.Rand, n int) []*ChangeSet {
 
 		var cs ChangeSet
 		for _, key := range keys {
-			cs.Pairs = append(cs.Pairs, items[key])
+			p := items[key]
+			cs.Pairs = append(cs.Pairs, &p)
 		}
 
-		changeSets = append(changeSets, &cs)
+		changeSets = append(changeSets, cs)
 	}
 	return changeSets
 }